// src/app/api/model/insuranceProduct.ts
/**
 * InsuranceAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Policy } from './policy';
import { Admin } from './admin';


export interface InsuranceProduct { 
    productId?: number;
    productName?: string | null;
    category?: string | null;
    description?: string | null;
    basePremium?: number;
    createdAt?: string | null;
    updatedAt?: string | null;
    createdBy?: number | null;
    createdByNavigation?: Admin;
    policies?: Array<Policy> | null;
}